requirements.txt
streamlit>=1.35.0
numpy>=1.24.0
plotly>=5.18.0
pandas>=2.0.0
anthropic>=0.25.0
openai>=1.0.0
groq>=0.4.0
google-generativeai>=0.5.0
